# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "KnowledgeEvidenceResponse": {
        "properties": {
          "anchor_status": {
            "default": "exact",
            "title": "Anchor Status",
            "type": "string"
          },
          "artifact_revision": {
            "title": "Artifact Revision",
            "type": "integer"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deep_link": {
            "title": "Deep Link",
            "type": "string"
          },
          "frozen_quote": {
            "title": "Frozen Quote",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "locator": {
            "additionalProperties": true,
            "title": "Locator",
            "type": "object"
          },
          "node_id": {
            "format": "uuid",
            "title": "Node Id",
            "type": "string"
          },
          "parse_artifact_id": {
            "format": "uuid",
            "title": "Parse Artifact Id",
            "type": "string"
          },
          "quote_hash": {
            "title": "Quote Hash",
            "type": "string"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "revision_number": {
            "title": "Revision Number",
            "type": "integer"
          },
          "section_id": {
            "format": "uuid",
            "title": "Section Id",
            "type": "string"
          },
          "section_ordinal": {
            "title": "Section Ordinal",
            "type": "integer"
          },
          "source_content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Content Hash"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "source_title": {
            "title": "Source Title",
            "type": "string"
          },
          "source_version_id": {
            "format": "uuid",
            "title": "Source Version Id",
            "type": "string"
          },
          "source_version_number": {
            "title": "Source Version Number",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "node_id",
          "revision_id",
          "revision_number",
          "source_id",
          "source_title",
          "source_version_id",
          "source_version_number",
          "source_content_hash",
          "parse_artifact_id",
          "artifact_revision",
          "section_id",
          "section_ordinal",
          "quote_hash",
          "content_hash",
          "locator",
          "frozen_quote",
          "deep_link",
          "created_at"
        ],
        "title": "KnowledgeEvidenceResponse",
        "type": "object"
      },
      "KnowledgeNodeCreate": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "default": "",
            "maxLength": 20000,
            "title": "Body",
            "type": "string"
          },
          "conditions": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "Conditions",
            "type": "array"
          },
          "exceptions": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "Exceptions",
            "type": "array"
          },
          "expected_version": {
            "exclusiveMinimum": 0.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/KnowledgeNodeKind"
          },
          "parent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "expected_version",
          "title"
        ],
        "title": "KnowledgeNodeCreate",
        "type": "object"
      },
      "KnowledgeNodeDelete": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "exclusiveMinimum": 0.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "KnowledgeNodeDelete",
        "type": "object"
      },
...
      "KnowledgeWriteResponse": {
        "properties": {
          "node_id": {
            "format": "uuid",
            "title": "Node Id",
            "type": "string"
          },
          "node_version": {
            "title": "Node Version",
            "type": "integer"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "result_id": {
            "format": "uuid",
            "title": "Result Id",
            "type": "string"
          },
          "snapshot": {
            "additionalProperties": true,
            "title": "Snapshot",
            "type": "object"
          }
        },
        "required": [
          "result_id",
          "request_id",
          "node_id",
          "node_version",
          "snapshot"
        ],
        "title": "KnowledgeWriteResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "D5 immutable source ingestion, parsing, structured AI extraction, and human candidate review API for the personal knowledge workbench.",
    "title": "Zixu Knowledge Workbench API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge-spaces/{space_id}/knowledge-evidence/{evidence_id}": {
      "get": {
        "operationId": "get_knowledge_evidence",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEvidenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Knowledge Evidence",
        "tags": [
          "knowledge-tree"
        ]
      }
    },
    "/api/v1/knowledge-spaces/{space_id}/knowledge-nodes": {
      "post": {
        "operationId": "create_knowledge_node",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeNodeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeWriteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Knowledge Node",
        "tags": [
          "knowledge-tree"
        ]
      }
    },
    "/api/v1/knowledge-spaces/{space_id}/knowledge-nodes/{node_id}": {
      "delete": {
        "operationId": "delete_knowledge_node",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeNodeDelete"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeWriteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Delete Knowledge Node",
        "tags": [
          "knowledge-tree"
        ]
      },
      "get": {
        "operationId": "get_knowledge_node",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeNodeDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Knowledge Node",
        "tags": [
          "knowledge-tree"
        ]
      },
      "patch": {
        "operationId": "edit_knowledge_node",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeNodeEdit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeWriteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Edit Knowledge Node",
        "tags": [
          "knowledge-tree"
        ]
      }
    },
    "/api/v1/knowledge-spaces/{space_id}/knowledge-nodes/{node_id}/move": {
      "post": {
        "operationId": "move_knowledge_node",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeNodeMove"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeWriteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Move Knowledge Node",
        "tags": [
          "knowledge-tree"
        ]
      }
    },
    "/api/v1/knowledge-spaces/{space_id}/knowledge-search": {
      "get": {
        "operationId": "search_knowledge_tree",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 500,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Search Knowledge Tree",
        "tags": [
          "knowledge-tree"
        ]
      }
    },
    "/api/v1/knowledge-spaces/{space_id}/knowledge-tree": {
      "get": {
        "operationId": "get_knowledge_tree",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeTreeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Knowledge Tree",
        "tags": [
          "knowledge-tree"
        ]
      }
    },
    "/api/v1/knowledge-spaces/{space_id}/knowledge-write-requests/{idempotency_key}": {
      "get": {
        "operationId": "get_knowledge_write_request",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "idempotency_key",
            "required": true,
            "schema": {
              "title": "Idempotency Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeWriteRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Knowledge Write Request",
        "tags": [
          "knowledge-tree"
        ]
      }
    },
...
  }
}